data_research/requirements.txt
pandas
psycopg2-binary
scipy
